no filenames without number
description: on date changes during measurement the first file of the day is named without a number. The next file gets the incremental number again.

As sometimes measurement series have to span two days, all files are named with a succeeding number.

This also fixes a bug when there are already named files present in the folder and they would have to be overridden.